Fix compilation errors about template with gcc and mingw. Fix a lot of warnings about standard c++11 compliance and about mismatching types comparisons (#193)

+++ src/examples/Applets/App_ModelViewer.cpp
@@ -166,7 +166,7 @@ public:
     for( std::set<Geometry*>::iterator it = geometries.begin(); it != geometries.end(); ++it )
     {
       total_draw_calls += (*it)->drawCalls().size();
-      for(int i=0; i < (*it)->drawCalls().size(); ++i )
+      for(size_t i=0; i < (*it)->drawCalls().size(); ++i )
         total_triangles += (*it)->drawCalls().at(i)->countTriangles();
     }
